rag_api/requirements.txt
fastapi==0.104.1
openai==1.3.7
qdrant-client==1.9.1
python-dotenv==1.0.0
cohere==5.5.3
uvicorn==0.24.0
pydantic==2.5.0